Tagging hotfix bsa_fix bsa_fix

+++ lib/pack_material/validations/matres_product_variant_schema.rb
@@ -16,8 +16,8 @@ module PackMaterialApp
     optional(:product_variant_number, :integer).filled(:int?)
     required(:old_product_code, Types::StrippedString).maybe(:str?)
     required(:supplier_lead_time, :integer).maybe(:int?)
-    required(:minimum_stock_level, :integer).maybe(:int?)
-    required(:re_order_stock_level, :integer).maybe(:int?)
+    required(:minimum_stock_level, :decimal).maybe(:decimal?)
+    required(:re_order_stock_level, :decimal).maybe(:decimal?)
     required(:use_fixed_batch_number, :bool).maybe(:bool?)
     required(:mr_internal_batch_number_id, :integer).maybe(:int?)
